Some basic refactorings
