+++ src/main/java/com/mercadopago/api/MercadoPagoClient.java
@@ -1,0 +1,7 @@
+package com.mercadopago.api;
+
+public interface MercadoPagoClient {
+
+	MercadoPagoToken retrieveNewTokenUsing(TokenClientCredentials clientCredentials);
+
+}
