+++ src/main/java/com/mercadopago/api/TokenClientCredentials.java
@@ -1,4 +1,4 @@
-package com.mercadopago;
+package com.mercadopago.api;
 
 public class TokenClientCredentials {
